# api-docs.yaml
openapi: 3.0.1
info:
  title: Erapulus API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
paths:
  /api/user:
    get:
      tags:
        - User
      summary: List users
      description: "List users by with filtering by university, type, and name"
      operationId: list-application-user
      parameters:
        - name: university
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationUserDto'
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
        "500":
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
  /api/user/{userId}:
    delete:
      tags:
        - User
      summary: Delete user
      description: Delete user by ID
      operationId: delete-user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: NO CONTENT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
        "404":
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
        "500":
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
...
components:
  schemas:
    ApplicationUserDto:
      required:
        - email
        - firstName
        - id
        - lastName
        - type
      type: object
      properties:
        id:
          type: integer
          format: int32
        type:
          type: string
          enum:
            - ADMINISTRATOR
            - UNIVERSITY_ADMINISTRATOR
            - EMPLOYEE
            - STUDENT
        firstName:
          type: string
        lastName:
          type: string
        universityId:
          type: integer
          format: int32
        email:
          type: string
    ServerResponse:
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT